Infinite precision works, sort of...
Minor catch, have to limit it to one level T_T basic structure of it is
clearly working, but when raised to a second level it screws with the
mandelbrot. It looks like the lower right corner sorta works, so I
suspect it might be an issue with positive/negative? Need more test
cases...